+++ src/lib.rs
@@ -399,6 +399,13 @@ impl<Conf: Config> LevelMasksExt for BitSet<Conf>{
     }
 }
 
+#[inline]
+fn data_block_start_index<Conf: Config>(level0_index: usize, level1_index: usize) -> usize{
+    let level0_offset = level0_index << (Conf::DataBitBlock::SIZE_POT_EXPONENT + Conf::Level1BitBlock::SIZE_POT_EXPONENT);
+    let level1_offset = level1_index << (Conf::DataBitBlock::SIZE_POT_EXPONENT);
+    level0_offset + level1_offset
+}
+
 #[derive(Clone, Debug)]
 pub struct DataBlock<Block>{
     pub start_index: usize,